fix: Trailing comments and whitespace
Before: trailing comments and whitespace caused lexer errors. If there
was a single expression with trailing comments or whitespace, the lexer
would not be "complete", but the parser would have already parsed a full
expression. So there would be unexpected characters.

After: Now, there's a `clean_up` function on the lexer that will gobble
up all trailing whitespace and comments (and discard them) after the
lexer steps to next token.